feat: parse done

+++ wasm/snapshot_parser/src/snapshot.rs
@@ -30,35 +30,35 @@ pub struct SnapshotMeta {
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct SnapshotInfo {
-    pub edge_count: u64,
+    pub edge_count: u32,
     pub meta: SnapshotMeta,
-    pub node_count: u64,
-    pub trace_function_count: u64,
+    pub node_count: u32,
+    pub trace_function_count: u32,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Snapshot {
-    pub edges: Vec<u64>,
-    pub locations: Vec<u64>,
-    pub nodes: Vec<u64>,
-    pub samples: Vec<u64>,
+    pub edges: Vec<u32>,
+    pub locations: Vec<u32>,
+    pub nodes: Vec<u32>,
+    pub samples: Vec<u32>,
     pub snapshot: SnapshotInfo,
     pub strings: Vec<String>,
-    pub trace_function_infos: Vec<u64>,
-    pub trace_tree: Vec<u64>,
+    pub trace_function_infos: Vec<u32>,
+    pub trace_tree: Vec<u32>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Node {
     pub node_type: String,
     pub name: String,
-    pub id: u64,
-    pub self_size: u64,
-    pub edge_count: u64,
-    pub trace_node_id: u64,
-    pub detachedness: u64,
-    pub from_edge_index: Vec<u64>,
-    pub to_edge_index: Vec<u64>,
+    pub id: u32,
+    pub self_size: u32,
+    pub edge_count: u32,
+    pub trace_node_id: u32,
+    pub detachedness: u32,
+    pub from_edge_index: Vec<u32>,
+    pub to_edge_index: Vec<u32>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -66,7 +66,7 @@ pub struct Edge {
     pub edge_type: String,
     pub name_or_index: String,
     pub to_node_index: usize,
-    pub to_node_id: u64,
+    pub to_node_id: u32,
     pub from_node_index: usize,
-    pub from_node_id: u64,
+    pub from_node_id: u32,
 }
